#57 Rename module (core → parsers)
