+++ pandas_dataclasses/__init__.py
@@ -2,7 +2,7 @@ __version__ = "0.3.0"
 
 
 # submodules
-from .core import *
+from .parsers import *
 from .specs import *
 from .typing import *
