+++ pandas_dataclasses/pandas/frame.py
@@ -13,7 +13,7 @@ from morecopy import copy
 
 
 # submodules
-from ..core import get_attrs, get_columns, get_data, get_factory, get_index
+from ..parsers import get_attrs, get_columns, get_data, get_factory, get_index
 from ..typing import DataClass, P, PandasClass
